Extract custom video player handling from LightboxConfig to LightboxController
